Initial FAQ commit
Started FAQ functionality by moving `on_message` and `faq` from
`ServerManagement` to the new `Faq` cog.
Added a markdown file with sample questions and responses.
