+++ assets/FAQ.md
@@ -1,0 +1,66 @@
+
+#Frequently Asked Questions
+
+####I received a letter informing me about receiving a C in one of my courses. What do I need to do?
+
+A letter grade of C has the potential of lowering your GPA to below 3.0. An overall GPA of 3.0 or higher is required in order to be able to graduate. Hence, this letter is only telling you that you have to make up for that C grade by either receiving an A in another course or by retaking the course you received a C in.
+
+
+####I received a letter telling me that I am on probation and warning me that I may get dismissed. What are the steps I need to take?
+
+An overall GPA of 3.0 or above is required in order to graduate. Since your overall GPA dropped below 3.0 and you have already taken nine (9) credit hours worth of classes, you received this letter to let you know that you have been put on probation. While on probation, you have to maintain a per-semester GPA of 3.0 or above or otherwise you will be dismissed from the program. As soon as you reach an overall GPA of 3.0 or above you will no longer be on probation.
+
+
+####I am on probation but are only taking courses with a Pass/Fail grade. Is that okay?
+
+Yes, a passing grade is considered to be above 3.0 and thus will not result in your dismissal assuming you pass that course. You will still remain on probation as your GPA will not change by taking a course with a Pass/Fail grade.
+
+
+####I plan to apply for reduced course load (RCL). What are the requirements?
+
+RCL is intended for students in their last semester as they may not have enough courses remaining for a full semester load.
+
+
+####I want to apply for CPT for an internship. What are the rules?
+
+In order to be eligible for curricular practical training (CPT), you have to be in good standing and taken courses at Wright State for at least two (2) semesters. The internship position has to be aligned with the the topics of the program you are in. Thus, typical applications for CPT include an offer letter for that position or some other form of summary of the duties to be performed.
+
+
+####My GPA dropped and I am loosing my scholarship. Where can I get a statement of support?
+
+The statement should outline the potential for improvement with respect to your grades. This statement can come from a faculty member that you have taken classes with who can speak on your behalf. The Graduate Program Director can provide a statement as well. However, the Graduate Program Director can only consider your grades as they appear on your transcript. Hence, there needs to be a clear trend of improvement in order to be able to write an effective statement for you.
+
+
+####I received a lower grade than I expected. What can I do?
+
+Faculty evaluate your work exactly as stated in the syllabus and combine the scores to translate into the final grade based on the rubric specified in the syllabus. A curve may be applied to improve the grades for all students. However, if a curve is applied and what this curve looks like is solely at the discretion of the intructor. Unless there are items you did not get scores for, there is little you can do.
+
+
+####I am in the masters or PhD program and received a D or F in a class. Do I have to retake the class?
+
+An F is a failing grade in any program and so is a D in any graduate program as it is not allowed to be applied toward a graduate degree. So yes, you will have to retake that class.
+
+
+####I am in the masters or PhD program and received a C in a class. Do I have to retake the class?
+
+Up to six (6) semester-hours of courses with a grade of C can be counted toward your graduate degree. However, a grade of C may push your GPA to below 3.0. Since you need a GPA of 3.0 or higher in order to graduate you will have to make up for any C grades you received by either receiving a grade of A in another course or retaking that class.
+
+
+####If I retake a class due to a low grade, how does that affect my GPA?
+
+In the graduate programs, you can replace up to two (2) grades entirely by retaking the class and finishing that class with a better grade. This will then result in the original grade not to be counted for your GPA. Beyond the two (2) allowed replacements, the new (and the old) grade will be included in the average GPA calculation just like any other grade.
+
+
+####What is the minimum number of credit hours I need to register as a graduate student?
+In order to be considered a full-time student at the graduate level, you need to register for at least six (6) credit hours. A half-time student only needs to register for three (3) credit hours. Additional information can be found here: https://www.wright.edu/raiderconnect/financial-aid/eligibility
+
+
+####I am an international student and will be arriving late for the semester. What do I need to do?
+This could be an issue if you will be arriving late for classes more than a week. Depending on the policy the instructors of your classes apply, you could run into problems there. So we stronly discourage late arrivals if at all possible. As an international student, you will have to fill out a form for late arrival with UCIE: https://www.wright.edu/international-education/international-students/forms
+
+####What are the language requirements for students whose native is not English?
+If your native language is not English and you do not have a degree from a regionally accredited U.S. college or university, you must have a minimum score on one of the following exams: a score of 79/120 on the Test of English as a Foreign Language (T.O.E.F.L.), a band 6 or higher through the International English Language Testing System (IELTS), or a  score of 57 or higher on the Pearson PTE. If you have graduated from a school in a country where English is the official language for higher education, you will not be required to demonstrate English proficiency.
+
+####What are the current tuition and fees at Wright State?
+You can find the most current information on tuition and fees on Wright State's web site: https://www.wright.edu/raiderconnect/accounts-and-bills/tuition-and-fees
+
